+++ clang/lib/Sema/SemaType.cpp
@@ -8413,10 +8413,8 @@ static void HandleNeonVectorTypeAttr(QualType &CurType, const ParsedAttr &Attr,
 /// Handle the __ptrauth qualifier.
 static void HandlePtrAuthQualifier(ASTContext &Ctx, QualType &T,
                                    const ParsedAttr &Attr, Sema &S) {
-  auto AttributeName = Attr.getAttrName()->getName();
   if (Attr.getNumArgs() < 1 || Attr.getNumArgs() > 3) {
-    S.Diag(Attr.getLoc(), diag::err_ptrauth_qualifier_bad_arg_count)
-        << AttributeName;
+    S.Diag(Attr.getLoc(), diag::err_ptrauth_qualifier_bad_arg_count);
     Attr.setInvalid();
     return;
   }